[Bridge Node] Make IP configurable
