+++ crates/sui-json-rpc/Cargo.toml
@@ -37,6 +37,7 @@ bcs.workspace = true
 eyre.workspace = true
 once_cell.workspace = true
 serde_json.workspace = true
+sui-network.workspace = true
 
 tap.workspace = true
