+++ crates/sui-network/src/lib.rs
@@ -2,7 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use mysten_network::config::Config;
+use std::net::{IpAddr, SocketAddr};
 use std::time::Duration;
+use tracing::error;
 
 pub mod api;
 pub mod discovery;
@@ -22,4 +24,16 @@ pub fn default_mysten_network_config() -> Config {
     net_config.request_timeout = Some(DEFAULT_REQUEST_TIMEOUT_SEC);
     net_config.http2_keepalive_interval = Some(DEFAULT_HTTP2_KEEPALIVE_SEC);
     net_config
+}
+
+pub fn parse_ip(ip: &str) -> Option<IpAddr> {
+    ip.parse::<IpAddr>().ok().or_else(|| {
+        ip.parse::<SocketAddr>()
+            .ok()
+            .map(|socket_addr| socket_addr.ip())
+            .or_else(|| {
+                error!("Failed to parse value of {:?} to ip address or socket.", ip,);
+                None
+            })
+    })
 }
